+++ src/package/themes/default/index.ts
@@ -2,6 +2,7 @@ import base from './base';
 import button from './components/button';
 import card from './components/card';
 import input from './components/input';
+import passwordInput from './components/password-input';
 import numberInput from './components/number-input';
 import formField from './components/form-field';
 import textarea from './components/textarea';
@@ -53,6 +54,7 @@ export default {
         button,
         card,
         input,
+        passwordInput,
         numberInput,
         formField,
         textarea,